requirements.txt
aiogram>=3.0.0b3
aioredis~=2.0
environs~=9.0
bs4
lxml
apschedule